+++ src/pylingdocs/output.py
@@ -422,8 +422,10 @@ class CLLD(OutputFormat):
         delim = "\n# "
         parts = tent.split(delim)[1::]
         if len(parts) == 0 or OUTPUT_TEMPLATES["clld"] in ["slides", "article"]:
+            # these use # as section markers, so we add a level for the html output
+            tent = tent.replace("\n#", "\n##")
             with open(my_output_dir / "content.txt", "w", encoding="utf-8") as f:
-                f.write(content)
+                f.write(tent)
         else:
             tag_dic = {}
             content_dic = {}