+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * bugs
 
 ### Changed
+* table metadata are stored in human-friendlier YAML
 
 ## [0.0.4] -- 2022-06-03